chore: refactor toward our standard
