+++ lms/djangoapps/verify_student/signals.py
@@ -9,7 +9,7 @@ from django.dispatch import Signal
 from django.dispatch.dispatcher import receiver
 from xmodule.modulestore.django import SignalHandler, modulestore
 
-from common.djangoapps.student.models_api import get_name, get_pending_name_change
+from common.djangoapps.student.api import get_name, get_pending_name_change
 from openedx.core.djangoapps.user_api.accounts.signals import USER_RETIRE_LMS_CRITICAL
 
 from .models import SoftwareSecurePhotoVerification, VerificationDeadline
